consolidate filter logic
